# docs/swagger.yaml
basePath: /api
definitions:
  api.CommonResponse:
    properties:
      data:
        type: object
      error:
        type: string
      status:
        type: integer
    type: object
  api.ModemShortInfo:
    properties:
      id:
        type: string
    type: object
info:
  contact: {}
  description: API for network management - AP, ethernet, modems.
  title: OrAP Service API
  version: "0.1"
paths:
  /auth/token:
    post:
      consumes:
      - application/json
      description: Get JWT token for sent username and password.
      produces:
      - application/json
      responses:
        "200":
          description: Common JSON response with JWT as data
          schema:
            allOf:
            - $ref: '#/definitions/api.CommonResponse'
            - properties:
                data:
                  type: string
              type: object
        "401":
          description: Unauthorized error
          schema:
            $ref: '#/definitions/api.CommonResponse'
        "500":
          description: Unhandled server error
          schema:
            $ref: '#/definitions/api.CommonResponse'
      security:
      - BasicAuth: []
      summary: User Login method
      tags:
      - auth
  /modems:
    get:
      consumes:
      - application/json
      description: 'Get list of modems with short info: id, name, description and
        connection state.'
      produces:
      - application/json
      responses:
        "200":
          description: Common JSON response with list of modems
          schema:
            allOf:
            - $ref: '#/definitions/api.CommonResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/api.ModemShortInfo'
                  type: array
              type: object
        "401":
          description: Unauthorized error
          schema:
            $ref: '#/definitions/api.CommonResponse'
        "500":
          description: Unhandled server error
          schema:
            $ref: '#/definitions/api.CommonResponse'
      security:
      - BasicAuth: []
      summary: List of modems
      tags:
      - modems_info
  /modems/{modem_id}:
    get:
      consumes:
      - application/json
      description: 'Get info about modem: common info, state, etc.'
      parameters:
      - description: Modem id from list of modems
        in: path
        name: modem_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Common JSON response with modem detailed info as data
          schema:
            allOf:
            - $ref: '#/definitions/api.CommonResponse'
            - properties:
                data:
                  type: string
              type: object
        "401":
          description: Unauthorized error
          schema:
            $ref: '#/definitions/api.CommonResponse'
        "404":
          description: Modem not found error
          schema:
            $ref: '#/definitions/api.CommonResponse'
        "500":
          description: Unhandled server error
          schema:
            $ref: '#/definitions/api.CommonResponse'
      security:
      - BasicAuth: []
      summary: Modem detailed information
      tags:
      - modems_info
securityDefinitions:
  BasicAuth:
    type: basic
swagger: "2.0"
tags:
- description: Auth operations
  name: auth
- description: Common operations
  name: common
- description: Modems info operations
  name: modems_info